test: consolidate flux alias tests into existing test files and remove redundant test file
